separate scripts and alias files for compounds:
