--- data/inputs/compound-test.txt
@@ -1,89 +1,0 @@
-54483521
-25033539
-9803433
-9549213
-23633604
-11180808
-981054
-6891506
-10430360
-3387354
-46926360
-15953870
-24988881
-9910486
-46944259
-6305587
-672245
-11950726
-24964624
-5280647
-44137946
-71077
-973957
-135565545
-9884685
-761523
-10341154
-2056026
-3091786
-669190
-45142457
-24748204
-1285940
-25019940
-5013
-1000
-10000009
-10000456
-100016
-100018
-10007
-10008718
-100094
-1001
-10011812
-10013505
-10016012
-10020353
-10021692
-10022508
-10022829
-10023700
-100252
-10026
-10026128
-10029385
-100299
-1003
-10033072
-10034073
-10039774
-1004
-10040286
-10041070
-100472
-108
-G00002CF
-G00009BX
-G00012MO
-G00012RZ
-G00013MO
-G00014MO
-G00015MO
-G00016MO
-G00017MO
-G00018MO
-G00019MO
-G00020MO
-G00021MO
-G00022MO
-G00025YC
-G00026MO
-G00027JG
-G00028MO
-G00029MO
-G00030HB
-G00030MO
-25098607
-G53621MK
