+++ data/inputs/compound_IDs_DrugCentral_test.txt
@@ -1,0 +1,100 @@
+1000
+10007
+10033072
+1004
+100472
+10071196
+10072
+10077130
+10080
+10083
+10089
+10090
+10096344
+10100
+10101
+10105
+10111431
+10113978
+10114
+10124
+10127622
+10130337
+10133
+10135765
+101534
+10178705
+10180
+10182969
+10184653
+10184665
+10198308
+102046
+10206
+10219
+10220503
+102210
+10235
+10239
+10240
+10250769
+10252
+10255
+102669
+10275
+10275777
+10280735
+10288191
+10290
+1030
+10304
+10306
+10307
+1031
+10311
+10311306
+10321047
+10324367
+10328936
+10347880
+10348973
+10351092
+10360683
+10374
+10391
+10404
+10406
+10444661
+10453870
+1046
+10472693
+104741
+104762
+104799
+10482134
+104838
+104845
+104850
+104865
+104903
+104926
+104976
+1050
+105024
+1051
+10517
+10518
+10520
+10531
+1054
+10548
+10577
+10599
+10606
+10607
+10610
+10612
+10616
+10617
+10620
+10630
