+++ data/inputs/compound_IDs_GlyTouCan_test.txt
@@ -1,0 +1,100 @@
+10000009
+10008718
+10011812
+10022829
+10023700
+10053824
+10062728
+10062742
+10062744
+10074840
+10077784
+10102976
+10103058
+10106330
+10121152
+10130220
+10137150
+10145029
+10145887
+10147060
+10147061
+10147853
+10148744
+10167273
+10167274
+10168639
+10169039
+10169040
+10170292
+10170667
+10177989
+10185565
+10187270
+10187357
+10188612
+10190843
+10197254
+10206531
+10212235
+10213173
+10213283
+102190
+10219674
+10220660
+10221549
+10221550
+102288
+10230776
+10230777
+10232075
+G99785MW
+G99792ZO
+G99797PH
+G99799ZQ
+G99801SM
+G99808VS
+G99811IO
+G99813OP
+G99814AM
+G99824ER
+G99831DA
+G99834WZ
+G99835YB
+G99839QA
+G99843HL
+G99844CS
+G99846HH
+G99848BA
+G99849MX
+G99849PP
+G99851EW
+G99869RW
+G99872BC
+G99873PN
+G99883ZA
+G99884YO
+G99891PR
+G99893CA
+G99897FQ
+G99902AI
+G99921WD
+G99923IA
+G99931NX
+G99932KY
+G99933ME
+G99939EJ
+G99941DT
+G99950SF
+G99952ON
+G99954IG
+G99966GV
+G99968DX
+G99975QE
+G99978QH
+G99980WG
+G99981ST
+G99984YM
+G99992LL
+G99994YV
+G99997WS
