+++ data/inputs/compound_IDs_LINCS.txt
@@ -1,0 +1,35 @@
+54483521
+25033539
+9803433
+9549213
+23633604
+11180808
+981054
+6891506
+10430360
+3387354
+46926360
+15953870
+24988881
+9910486
+46944259
+6305587
+672245
+11950726
+24964624
+5280647
+44137946
+71077
+973957
+135565545
+9884685
+761523
+10341154
+2056026
+3091786
+669190
+45142457
+24748204
+1285940
+25019940
+5013
