+++ data/inputs/compound_IDs_test.txt
@@ -1,0 +1,135 @@
+1000
+1004
+1030
+1031
+1046
+1050
+1051
+1054
+5013
+10007
+10072
+10080
+10083
+10089
+10090
+10100
+10101
+10105
+10114
+10124
+10133
+10180
+10206
+10219
+10235
+10239
+10240
+10252
+10255
+10275
+10290
+10304
+10306
+10307
+10311
+10374
+10391
+10404
+10406
+10517
+10518
+10520
+10531
+10548
+10577
+10599
+10606
+10607
+10610
+10612
+10616
+10617
+10620
+10630
+71077
+100472
+101534
+102046
+102210
+102669
+104741
+104762
+104799
+104838
+104845
+104850
+104865
+104903
+104926
+104976
+105024
+669190
+672245
+761523
+973957
+981054
+1285940
+2056026
+3091786
+3387354
+5280647
+6305587
+6891506
+9549213
+9803433
+9884685
+9910486
+10033072
+10071196
+10077130
+10096344
+10111431
+10113978
+10127622
+10130337
+10135765
+10178705
+10182969
+10184653
+10184665
+10198308
+10220503
+10250769
+10275777
+10280735
+10288191
+10311306
+10321047
+10324367
+10328936
+10341154
+10347880
+10348973
+10351092
+10360683
+10430360
+10444661
+10453870
+10472693
+10482134
+11180808
+11950726
+15953870
+23633604
+24748204
+24964624
+24988881
+25019940
+25033539
+44137946
+45142457
+46926360
+46944259
+54483521
+135565545
